I have done changed to .env too, please review.
