+++ environment.env
@@ -1,3 +1,6 @@
+MONGO_DBNAME=SE-2PM2
+MONGO_URI ="mongodb://yl7408:tR%@Vw$5#Y6@r$c@se-2pm2.xbstfiu.mongodb.net/?retryWrites=true&w=majority";
 FLASK_APP=app.py
 FLASK_ENV=development
 FLASK_PORT=5000
+GITHUB_REPO = https://github.com/yl7408
