+++ index.html
@@ -1,8 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-    <script>
-        const mongoURI = "mongodb+srv://yl7408:tR%@Vw$5#Y6@r$c@se-2pm2.xbstfiu.mongodb.net/?retryWrites=true&w=majority";
-    </script>
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
